Trove kilo-rc1 milestone (2015.1.0rc1)

+++ trove/guestagent/datastore/mysql/service.py
@@ -58,15 +58,16 @@ INCLUDE_MARKER_OPERATORS = {
 }
 
 OS_NAME = operating_system.get_os()
-MYSQL_CONFIG = {operating_system.REDHAT: "/etc/mysql/my.cnf",
+MYSQL_CONFIG = {operating_system.REDHAT: "/etc/my.cnf",
                 operating_system.DEBIAN: "/etc/mysql/my.cnf",
                 operating_system.SUSE: "/etc/my.cnf"}[OS_NAME]
 MYSQL_SERVICE_CANDIDATES = ["mysql", "mysqld", "mysql-server"]
 MYSQL_BIN_CANDIDATES = ["/usr/sbin/mysqld", "/usr/libexec/mysqld"]
 MYCNF_OVERRIDES = "/etc/mysql/conf.d/overrides.cnf"
 MYCNF_OVERRIDES_TMP = "/tmp/overrides.cnf.tmp"
-MYCNF_REPLMASTER = "/etc/mysql/conf.d/0replication.cnf"
-MYCNF_REPLMASTER_TMP = "/tmp/replication.cnf.tmp"
+MYCNF_REPLMASTER = "/etc/mysql/conf.d/0replmaster.cnf"
+MYCNF_REPLSLAVE = "/etc/mysql/conf.d/1replslave.cnf"
+MYCNF_REPLCONFIG_TMP = "/tmp/replication.cnf.tmp"
 
 
 # Create a package impl
@@ -421,6 +422,9 @@ class MySqlAdmin(object):
     def list_databases(self, limit=None, marker=None, include_marker=False):
         """List databases the user created on this mysql instance."""
         LOG.debug("---Listing Databases---")
+        ignored_database_names = "'%s'" % "', '".join(CONF.ignore_dbs)
+        LOG.debug("The following database names are on ignore list and will "
+                  "be omitted from the listing: %s" % ignored_database_names)
         databases = []
         with LocalSqlClient(get_engine()) as client:
             # If you have an external volume mounted at /var/lib/mysql
@@ -434,10 +438,7 @@ class MySqlAdmin(object):
                 'default_collation_name as collation',
             ]
             q.tables = ['information_schema.schemata']
-            q.where = ["schema_name NOT IN ("
-                       "'mysql', 'information_schema', "
-                       "'lost+found', '#mysql50#lost+found'"
-                       ")"]
+            q.where = ["schema_name NOT IN (" + ignored_database_names + ")"]
             q.order = ['schema_name ASC']
             if limit:
                 q.limit = limit + 1
@@ -738,6 +739,11 @@ class MySqlApp(object):
                     LOG.exception(_("Unable to set %(key)s with value "
                                     "%(value)s.") % output)
 
+    def make_read_only(self, read_only):
+        with LocalSqlClient(get_engine()) as client:
+            q = "set global read_only = %s" % read_only
+            client.execute(text(str(q)))
+
     def _write_temp_mycnf_with_admin_account(self, original_file_path,
                                              temp_file_path, password):
         mycnf_file = open(original_file_path, 'r')
@@ -822,26 +828,42 @@ class MySqlApp(object):
         if os.path.exists(MYCNF_OVERRIDES):
             utils.execute_with_timeout("sudo", "rm", MYCNF_OVERRIDES)
 
-    def write_replication_overrides(self, overrideValues):
+    def _write_replication_overrides(self, overrideValues, cnf_file):
         LOG.info(_("Writing replication.cnf file."))
 
-        with open(MYCNF_REPLMASTER_TMP, 'w') as overrides:
+        with open(MYCNF_REPLCONFIG_TMP, 'w') as overrides:
             overrides.write(overrideValues)
         LOG.debug("Moving temp replication.cnf into correct location.")
-        utils.execute_with_timeout("sudo", "mv", MYCNF_REPLMASTER_TMP,
-                                   MYCNF_REPLMASTER)
+        utils.execute_with_timeout("sudo", "mv", MYCNF_REPLCONFIG_TMP,
+                                   cnf_file)
 
         LOG.debug("Setting permissions on replication.cnf.")
-        utils.execute_with_timeout("sudo", "chmod", "0644",
-                                   MYCNF_REPLMASTER)
+        utils.execute_with_timeout("sudo", "chmod", "0644", cnf_file)
 
-    def remove_replication_overrides(self):
+    def _remove_replication_overrides(self, cnf_file):
         LOG.info(_("Removing replication configuration file."))
-        if os.path.exists(MYCNF_REPLMASTER):
-            utils.execute_with_timeout("sudo", "rm", MYCNF_REPLMASTER)
+        if os.path.exists(cnf_file):
+            utils.execute_with_timeout("sudo", "rm", cnf_file)
+
+    def exists_replication_source_overrides(self):
+        return os.path.exists(MYCNF_REPLMASTER)
+
+    def write_replication_source_overrides(self, overrideValues):
+        self._write_replication_overrides(overrideValues, MYCNF_REPLMASTER)
+
+    def write_replication_replica_overrides(self, overrideValues):
+        self._write_replication_overrides(overrideValues, MYCNF_REPLSLAVE)
+
+    def remove_replication_source_overrides(self):
+        self._remove_replication_overrides(MYCNF_REPLMASTER)
+
+    def remove_replication_replica_overrides(self):
+        self._remove_replication_overrides(MYCNF_REPLSLAVE)
 
     def grant_replication_privilege(self, replication_user):
         LOG.info(_("Granting Replication Slave privilege."))
+
+        LOG.debug("grant_replication_privilege: %s" % replication_user)
 
         with LocalSqlClient(get_engine()) as client:
             g = sql_query.Grant(permissions=['REPLICATION SLAVE'],
@@ -849,19 +871,6 @@ class MySqlApp(object):
                                 clear=replication_user['password'])
 
             t = text(str(g))
-            client.execute(t)
-
-    def revoke_replication_privilege(self):
-        LOG.info(_("Revoking Replication Slave privilege."))
-
-        with LocalSqlClient(get_engine()) as client:
-            results = client.execute('SHOW SLAVE STATUS').fetchall()
-            slave_status_info = results[0]
-
-            r = sql_query.Revoke(permissions=['REPLICATION SLAVE'],
-                                 user=slave_status_info['master_user'])
-
-            t = text(str(r))
             client.execute(t)
 
     def get_port(self):
@@ -878,27 +887,10 @@ class MySqlApp(object):
             }
             return binlog_position
 
-    def change_master_for_binlog(self, host, port, logging_config):
-        LOG.info(_("Configuring replication from %s.") % host)
-
-        replication_user = logging_config['replication_user']
-        change_master_cmd = ("CHANGE MASTER TO MASTER_HOST='%(host)s', "
-                             "MASTER_PORT=%(port)s, "
-                             "MASTER_USER='%(user)s', "
-                             "MASTER_PASSWORD='%(password)s', "
-                             "MASTER_LOG_FILE='%(log_file)s', "
-                             "MASTER_LOG_POS=%(log_pos)s" %
-                             {
-                                 'host': host,
-                                 'port': port,
-                                 'user': replication_user['name'],
-                                 'password': replication_user['password'],
-                                 'log_file': logging_config['log_file'],
-                                 'log_pos': logging_config['log_position']
-                             })
-
+    def execute_on_client(self, sql_statement):
+        LOG.debug("Executing SQL: %s" % sql_statement)
         with LocalSqlClient(get_engine()) as client:
-            client.execute(change_master_cmd)
+            return client.execute(sql_statement)
 
     def start_slave(self):
         LOG.info(_("Starting slave replication."))
@@ -906,7 +898,7 @@ class MySqlApp(object):
             client.execute('START SLAVE')
             self._wait_for_slave_status("ON", client, 60)
 
-    def stop_slave(self):
+    def stop_slave(self, for_failover):
         replication_user = None
         LOG.info(_("Stopping slave replication."))
         with LocalSqlClient(get_engine()) as client:
@@ -915,10 +907,16 @@ class MySqlApp(object):
             client.execute('STOP SLAVE')
             client.execute('RESET SLAVE ALL')
             self._wait_for_slave_status("OFF", client, 30)
-            client.execute('DROP USER ' + replication_user)
+            if not for_failover:
+                client.execute('DROP USER ' + replication_user)
         return {
             'replication_user': replication_user
         }
+
+    def stop_master(self):
+        LOG.info(_("Stopping replication master."))
+        with LocalSqlClient(get_engine()) as client:
+            client.execute('RESET MASTER')
 
     def _wait_for_slave_status(self, status, client, max_time):
 
@@ -988,6 +986,54 @@ class MySqlApp(object):
         config_contents = configuration['config_contents']
         LOG.info(_("Resetting configuration."))
         self._write_mycnf(None, config_contents)
+
+    # DEPRECATED: Mantain for API Compatibility
+    def get_txn_count(self):
+        LOG.info(_("Retrieving latest txn id."))
+        txn_count = 0
+        with LocalSqlClient(get_engine()) as client:
+            result = client.execute('SELECT @@global.gtid_executed').first()
+            for uuid_set in result[0].split(','):
+                for interval in uuid_set.split(':')[1:]:
+                    if '-' in interval:
+                        iparts = interval.split('-')
+                        txn_count += int(iparts[1]) - int(iparts[0])
+                    else:
+                        txn_count += 1
+        return txn_count
+
+    def _get_slave_status(self):
+        with LocalSqlClient(get_engine()) as client:
+            return client.execute('SHOW SLAVE STATUS').first()
+
+    def _get_master_UUID(self):
+        slave_status = self._get_slave_status()
+        return slave_status and slave_status['Master_UUID'] or None
+
+    def _get_gtid_executed(self):
+        with LocalSqlClient(get_engine()) as client:
+            return client.execute('SELECT @@global.gtid_executed').first()[0]
+
+    def get_last_txn(self):
+        master_UUID = self._get_master_UUID()
+        last_txn_id = '0'
+        gtid_executed = self._get_gtid_executed()
+        for gtid_set in gtid_executed.split(','):
+            uuid_set = gtid_set.split(':')
+            if uuid_set[0] == master_UUID:
+                last_txn_id = uuid_set[-1].split('-')[-1]
+                break
+        return master_UUID, int(last_txn_id)
+
+    def get_latest_txn_id(self):
+        LOG.info(_("Retrieving latest txn id."))
+        return self._get_gtid_executed()
+
+    def wait_for_txn(self, txn):
+        LOG.info(_("Waiting on txn '%s'.") % txn)
+        with LocalSqlClient(get_engine()) as client:
+            client.execute("SELECT WAIT_UNTIL_SQL_THREAD_AFTER_GTIDS('%s')"
+                           % txn)
 
 
 class MySqlRootAccess(object):
